requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic[email]==2.10.3
supabase>=2.10.0,<3
httpx==0.27.2
